Feature: Sistema de Cura e outras atualizações na estrutura do código
