+++ src/Clerigo.java
@@ -1,7 +1,7 @@
 import java.util.Random;
 
-public class Arqueiro extends Personagem {
-    public Arqueiro(String nome, String arma, int vida, int forca, int defesa, int destreza, int estusFlask) {
+public class Clerigo extends Personagem {
+    public Clerigo(String nome, String arma, int vida, int forca, int defesa, int destreza, int estusFlask) {
         super(nome, arma, vida, forca, defesa, destreza, estusFlask);
     }
 
@@ -11,7 +11,7 @@ public class Arqueiro extends Personagem {
         int min = getForca() / 2;
         int max = getForca() + getDefesa();
         int dano = rand.nextInt(max - min + 1) + min;
-        System.out.printf("Você atirou uma flecha com o seu %s e causou %d de dano!\n", getArma(), dano);
+        System.out.printf("Você atacou com o seu %s e causou %d de dano! \n", getArma(), dano);
         return dano;
     }
